s21_Python/Python_Bootcamp.Day6/src/ex02/requirements.txt
protobuf
grpcio
grpcio-tools
pydantic
psycopg2-binary